Remove ActivityListViewModelTests
